msm: camera: actuator: remove unneeded copy_from_user operation for init_settings
the allocation of enough space for init_settings is sometimes failing
due to a mysterious bug in memory managment.
init_settings is only being read in the succeeding functions, and
copying a read-only structure from user to kernel holds little sense.
Avoid the allocation bug by simply not copying the structure over to
kernel and accessing the information from the user structure directly.
the only function that's called after the allocation is
msm_actuator_init_focus().
the only thing that  msm_actuator_init_focus() does with init_settings
is either an i2c_write() or an i2c_poll().
as far as I could read, both functions copy the data.
after the function returns, init_settings is deallocated, so the
structure is not used after this point.
the data structure init_settings remains immuted during the whole
function.