ldo_vibrator: Implement intensity control
NOTE : THIS COMMIT IS IN TESTING STATE, DO NOT MERGE YET!

I have re-worked a fair bit of the LDO vibrator driver to
support intensity adjustment. This is achieved by simulating a
PWM.

The intensity regulator only adjusts vibrstion times in
microseconds ranging from MIN_TRG_ADJUST_VAL to MAX_TRG_ADJUST_VAL

The intensity of the vibrations is regulated using the following
preprocessors :

LDO_VIBRATOR_INTENSITY_HIGH	- High duty cycles in microseconds
LDO_VIBRATOR_INTENSITY_LOW	- Low  duty cycles in microseconds

Both values depend on each other, however as for now I obtained
best results using a 25% LOW of the set HIGH value

To visulize changes graphically between the old method and the
current one

100ms vibration on old driver

V
|   -------------
|  /
| /
|/
____________________100ms t

As you can see the vibration motor gains velocity on the
beginning, then reaches it's max point. While the first couple
of values are acceptable in intensity, once the max point has
reached, the vibrations become unconfortably strong,
unfortunately this may only be controlled with voltage, and
because this doesn't work on digital signals, you're pretty much
stuck

100ms vibration method 2
V
|
|-------------------- Avrg intensity
| /\  /\  /\  /\  /\
|/  \/  \/  \/  \/  \
_____________________100ms t

As you can see, we're not really magically changing the output
voltage or the vibrator, instead we're constantly accelerating
it for a very short time. The point of descending is then our
targeted intensity. The down time then decides smoothness and
must be adjusted appropriately

However there are a few things to note, which marks this driver
change as testing

Making the HIGH value too low will result in a heavy perfomance
loss during multiple vibration requests at once. The reason for
this is most likely usleep related. Solutions are planned for
this issue. Another thing to note is that this hasn't been
written with the most kernel like method, meaning this code may
be considered as unprofessional or not high enough for usual
kernel standarts.

Change-Id: I95e4c963a3f846bbe356a9aecb2beb1597944444
Signed-off-by: CTXz <ctx.xda@gmail.com>